+++ MangoPay/PayOutEligibilityResponse.php
@@ -10,4 +10,12 @@ class PayOutEligibilityResponse extends Dto
      * @var InstantPayout
      */
     public $InstantPayout;
+
+    public function GetSubObjects()
+    {
+        $subObjects = parent::GetSubObjects();
+        $subObjects['InstantPayout'] = '\MangoPay\InstantPayout';
+
+        return $subObjects;
+    }
 }